### 2015-05-21 #14 @uribo
lab.note
* 1. What does this package do? (explain in 50 words or less)
Create laboratory note and report for helpful reproducible research.
* 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

```
Package: lab.note
Title: Create Laboratory Note and Presentation Using R Markdown
Version: 0.7.0
Authors@R: person("Shinya", "Uryu", email = "suika1127@gmail.com", role = c("aut", "cre"))
  broom,
  ggvis,
  knitcitations,
  shiny,
  xtable
Description: Make it can reproducible R Markdown report and other format such as presentation and RPubs published files. Contains built-in useful package using R and LaTeX. These packages usages to see vignettes and wiki more information.
```
* 3. URL for the package (the development repository, not a stylized html page)
    * https://github.com/uribo/lab.note
* 4. What data source(s) does it work with (if applicable)?
    * none.
* 5. Who is the target audience?
    * academic student and researcher.
* 6. Check the box next to each policy below, confirming that you agree. These are mandatory.
* [x] This package does not violate the Terms of Service of any service it interacts with.
* [x] The repository has continuous integration with Travis and/or another service
* [x] The package contains a vignette
* [x] The package contains a reasonably complete readme with devtools install instructions
* [ ] The package contains unit tests
* 7. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
* [x] Are there any package dependencies not on CRAN?
* [x] Does the package have a CRAN accepted license?
* [x] Did `devtools::check()` produce any errors or warnings? If so paste them below.
    * no
* 8. Please add explanations below for any exceptions to the above:
* 9. If this is a resubmission following rejection, please explain the change in cirucmstances.
