### 2015-05-21 #14 @sckott
hi @uribo thanks for the submission. It looks like you do use a CRAN accepted license (MIT) - please check that box. Also, answer the questions (if answer is No, then write that.)

- Are there any package dependencies not on CRAN?
- Did devtools::check() produce any errors or warnings?
