### 2015-05-21 #14 @sckott
@cboettig assigned
