### 2015-05-22 #14 @uribo
Sorry, @sckott

I'm fixed checkbox and answered for question.

* Are there other R packages that accomplish the same thing? If so, what is different about yours?

similar packages (Rmd template packages)
*  mosaic ([CRAN](http://cran.r-project.org/web/packages/mosaic/index.html))
* rticles ([GitHub](https://github.com/rstudio/rticles))
* rmdTemplates ([GitHub](https://github.com/Pakillo/rmdTemplates))
* MarkdownMS ([GitHub](https://github.com/MarcoDVisser/MarkdownMS))

`lab.note` contains helpful and useful packages (i.e. knitr chunk option, custom font) using R reporting for **LaTeX beginner**. Also, default RMarkdown beamer templates are difficult to use while `lab.note` beamer template is already some initial configuration.
