+++ inst/rmarkdown/templates/literature/skeleton/references.bib
@@ -1,0 +1,44 @@
+@Manual{xie2015knitr,
+  title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
+  author = {Yihui Xie},
+  year = {2015},
+  note = {R package version 1.10.5},
+  url = {http://yihui.name/knitr/},
+}
+
+@Book{xie2013dynamic,
+  title = {Dynamic Documents with {R} and knitr},
+  author = {Yihui Xie},
+  publisher = {Chapman and Hall/CRC},
+  address = {Boca Raton, Florida},
+  year = {2013},
+  note = {ISBN 978-1482203530},
+  url = {http://yihui.name/knitr/},
+}
+
+@InCollection{xie2014knitr,
+  booktitle = {Implementing Reproducible Computational Research},
+  editor = {Victoria Stodden and Friedrich Leisch and Roger D. Peng},
+  title = {knitr: A Comprehensive Tool for Reproducible Research in {R}},
+  author = {Yihui Xie},
+  publisher = {Chapman and Hall/CRC},
+  year = {2014},
+  note = {ISBN 978-1466561595},
+  url = {http://www.crcpress.com/product/isbn/9781466561595},
+}
+
+@Manual{boettiger2015knitcitations,
+  title = {knitcitations: Citations for Knitr Markdown Files},
+  author = {Carl Boettiger},
+  year = {2015},
+  note = {R package version 1.0.6},
+  url = {http://CRAN.R-project.org/package=knitcitations},
+}
+
+@Manual{allaire2015rmarkdown,
+  title = {rmarkdown: Dynamic Documents for R},
+  author = {JJ Allaire and Joe Cheng and Yihui Xie and Jonathan McPherson and Winston Chang and Jeff Allen and Hadley Wickham and Rob Hyndman},
+  year = {2015},
+  note = {R package version 0.6.1},
+  url = {http://CRAN.R-project.org/package=rmarkdown},
+}
