+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: lab.note
 Type: Package
 Title: Create Laboratory Note and Presentation Using R Markdown
-Version: 0.7.0
-Date: 2015-05-21
+Version: 0.8.0
+Date: 2015-05-29
 Description: Make it can reproducible R Markdown report and other format 
   such as presentation and RPubs published files. Contains built-in useful 
   package using R and LaTeX. These packages usages to see vignettes and wiki 
