### 2015-05-31 a6187b0
[CLEAN]remove unused line
