### 2015-05-31 6a90bee
[UPGRADE]bump version for 0.8.0 :rocket:
