+++ inst/rmarkdown/templates/rpubs/skeleton/skeleton.Rmd
@@ -1,9 +1,12 @@
 ---
-# THEME: RPubs by lab.note ver. 0.8.0
+# THEME: RPubs by lab.note ver. 0.8.1
 title: "Change the title"
 author: "your names"
 date: "`r format(Sys.time(), '%B %d, %Y')`"
-output: html_document
+output: 
+  html_document:
+    theme: readable
+    toc: false
 ---
 
 ```{r [SETTING], echo = TRUE, message = FALSE}
