+++ inst/rmarkdown/templates/report/skeleton/skeleton.Rmd
@@ -1,5 +1,5 @@
 ---
-# THEME: Report by lab.note ver. 0.8.0
+# THEME: Report by lab.note ver. 0.8.1
 title: "Change the title"
 author: "Shinya Uryu"
 date: "`r format(Sys.time(), '%B %d, %Y')`"
