+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: lab.note
 Type: Package
 Title: Create Laboratory Note and Presentation Using R Markdown
-Version: 0.8.0
+Version: 0.8.1
 Date: 2015-05-29
 Description: Make it can reproducible R Markdown report and other format 
   such as presentation and RPubs published files. Contains built-in useful 
@@ -11,9 +11,9 @@ Authors@R: person("Shinya", "Uryu", email = "suika1127@gmail.com", role = c("aut
 Imports:
   knitr,
   rmarkdown
-Suggests: 
+Suggests:
   broom,
-  dplyr, 
+  dplyr,
   ggplot2,
   ggvis,
   knitcitations,
@@ -21,7 +21,7 @@ Suggests:
   shiny,
   tidyr,
   xtable
-VignetteBuilder: 
+VignetteBuilder:
   knitr
 URL: https://github.com/uribo/lab.note
 BugReports: https://github.com/uribo/lab.note/issues
