+++ inst/rmarkdown/templates/seminar/skeleton/skeleton.Rmd
@@ -6,29 +6,29 @@ author: \faTwitter @u_ribo
 date: "`r format(Sys.time(), '%B %d, %Y')`"
 header-includes:
 # define macro
-    - \definecolor{Black1}{RGB}{43,40,40}
-    - \definecolor{White1}{RGB}{255, 255, 250}
-    - \definecolor{Blue1}{RGB}{48, 122, 190} \definecolor{Blue2}{RGB}{99, 151, 205}
-    - \definecolor{Grey1}{RGB}{164, 173, 185}
+  - \definecolor{Black1}{RGB}{43,40,40}
+  - \definecolor{White1}{RGB}{255, 255, 250}
+  - \definecolor{Blue1}{RGB}{48, 122, 190} \definecolor{Blue2}{RGB}{99, 151, 205}
+  - \definecolor{Grey1}{RGB}{164, 173, 185}
 # using package, setup
-    - \usepackage{natbib}
-#    - \usepackage{zxjatype}
-#    - \setjamainfont{Hiragino Kaku Gothic Pro}
-    - \usepackage{fontspec, fontawesome} # require installed FontAwesome
+  - \usepackage{natbib}
+#  - \usepackage{zxjatype}
+#  - \setjamainfont{Hiragino Kaku Gothic Pro}
+  - \usepackage{fontspec, fontawesome} # require installed FontAwesome
 ## comment out when if you using image for slide background 
-#   - \usebackgroundtemplate{\includegraphics[width = \paperwidth, height = \paperheight]{...png}}
+#  - \usebackgroundtemplate{\includegraphics[width = \paperwidth, height = \paperheight]{...png}}
 # set beamer theme
-    - \setbeamercolor{background canvas}{bg = White1}
-    - \setbeamertemplate{navigation symbols}{}
-    - \setbeamertemplate{itemize items}{\textcolor{Blue2}{\faCaretRight}}
-    - \setbeamerfont{title}{size = \fontsize{38}{10}} \setbeamercolor{title}{fg = Blue1}
-    - \setbeamerfont{subtitle}{size = \large} \setbeamercolor{subtitle}{fg = Blue2}
-    - \setbeamercolor{author}{fg = Black1}
-    - \setbeamercolor{normal text}{fg = Black1}
-    - \setbeamerfont{date}{series = \itshape} \setbeamercolor{date}{fg = Grey1}
-    - \setbeamercolor{frametitle}{fg = Blue1}
+  - \setbeamercolor{background canvas}{bg = White1}
+  - \setbeamertemplate{navigation symbols}{}
+  - \setbeamertemplate{itemize items}{\textcolor{Blue2}{\faCaretRight}}
+  - \setbeamerfont{title}{size = \fontsize{38}{10}} \setbeamercolor{title}{fg = Blue1}
+  - \setbeamerfont{subtitle}{size = \large} \setbeamercolor{subtitle}{fg = Blue2}
+  - \setbeamercolor{author}{fg = Black1}
+  - \setbeamercolor{normal text}{fg = Black1}
+  - \setbeamerfont{date}{series = \itshape} \setbeamercolor{date}{fg = Grey1}
+  - \setbeamercolor{frametitle}{fg = Blue1}
 # new command
-    - \renewcommand{\baselinestretch}{1.2}
+  - \renewcommand{\baselinestretch}{1.2}
 output:
   beamer_presentation:
     fonttheme: "structurebold"
