### 2015-06-25 a4782ca
[MODIFY]chunk option: `tidy = FALSE`
