+++ inst/rmarkdown/templates/report/skeleton/skeleton.Rmd
@@ -39,7 +39,7 @@ options(width = 90, xtable.size = "footnotesize",
         xtable.include.rownames = FALSE, xtable.NA.string         = "NA")
 ### knitr chunk options ###
 opts_chunk$set(chace    = FALSE, echo      = FALSE, message  = FALSE, warning   =  FALSE,
-               tidy     = TRUE,  tidy.opts = list(width.cutoff = 70),
+               tidy     = FALSE, tidy.opts = list(width.cutoff = 70),
                fig.path = "Figure/", fig.width = 6, fig.height = 4, fig.retina = 1,
                dev      = c("pdf", "png"), dpi      = 150)
 theme_set(theme_classic(base_size = 18, base_family = "Helvetica"))
