+++ inst/rmarkdown/templates/lightning_talk/skeleton/document_setting.Rmd
@@ -12,7 +12,7 @@ opts_chunk$set(
   echo       = FALSE,
   warning    = FALSE,
   message    = FALSE,
-  tidy       = TRUE,
+  tidy       = FALSE,
   size       = "footnotesize",
   fig.path   = "images/",
   fig.width  = 8, 
