+++ inst/rmarkdown/templates/seminar/skeleton/document_setting.Rmd
@@ -12,8 +12,8 @@ opts_chunk$set(
   echo       = FALSE,
   warning    = FALSE,
   message    = FALSE,
-  tidy       = TRUE,
-  size       = 'large',
+  tidy       = FALSE,
+  size       = "large",
   fig.path   = "images/",
   fig.width  = 8, 
   fig.height = 8,
