### 2015-06-29 #14 @sckott
Hi @uribo - Sorry for the delay.

We discussed your submission, and have decided that your submission is out of scope for rOpenSci.  We appreciate the work that has been done, and wish you the best!
